input: update all taproot script spends to optionally make the ctrl block
In this commit, we update all the taproot scripts spends to optionally make the control block. This is useful in cases where we've already created the control block, or may not have the items needed to construct it in the first place.

We also add the control block to the sign descriptor itself.